Add special case code to work around syscall default behavior
New functionality pulled from falcosecurity/libs, introduces the
ability for the scap_open_* callers to specify initial system calls
of interest; if none are specified, default behavior is supposed to
match revert to 'capture all system calls'.  But due to a bug in the
falcosecurity/libs code, the 'default' behavior omits the capturing
of some system calls, including execve.  This change implements a
workaround to the bug, forcing the correct default.