Fix parse_execve_exit EXECVEAT path, to pass the current event to parse_dirfd()
New EXECVEAT-specific logic in parse_execve_exit() incorrectly passes
the incomplete matching enter_evt to parse_dirfd().  This does not
match normal usage, and is clearly incorrect -- the job of the parser
is to build up the state of the current event and its threadinfo, not
the state of the partially-populated, and soon to be defunct, entry_evt.