fix: thread-safety issue in proclist savefile
Since the proclist is not thread-safe, we had some corrupted captures
because the length was pre-computed and then when the actual item was
written it was different, hence the total was off.

Reduce the likelihood of this by using a secondary dumper, summing up
the lengths in a single loop and then writing everything at once.

Signed-off-by: Angelo Puglisi <angelopuglisi86@gmail.com>