new(userspace): consider scap host root for passwd and group lookup
Take users and groups considering scap host root.
When no host root is set, fall back using `getpwent`, `getgrent`, `getpwuid`, and `getgrgid`.

This way, when running on the host we will preserve the lookup with NSS,
when running as a container we will show the host users and groups
instead of the container ones (from files).

This applies to scap_create_userlist and sinsp_usergroup_manager

Signed-off-by: Angelo Puglisi <angelopuglisi86@gmail.com>