Include field name + cmp + num values in hands again
The prior version used static names for all filterchecks. This adds
back the ability to use the field name + cmp + num values without
incurring the overhead of hash lookups on push/pop. The hand is found
when the filtercheck is created instead.

Also remove the rule_owner stuff, not needed anymore with a stack.