new(scap,gvisor): configurable epoll_wait timeout
gvisor's next() calls epoll_wait to wait for the next event,
with an infinite timeout. Other engines don't block in next()
forever, so to remain compatible (and allow other things in
the main thread except for a scap_next() loop), make the timeout
configurable (still defaults to infinite at the sinsp layer).

Signed-off-by: Grzegorz Nosek <grzegorz.nosek@sysdig.com>