new(scap): Support cgroup namespaces hack
With cgroup namespaces enabled, all /proc/<pid>/cgroup entries
are relative to the namespace root, so we need a way to get
the absolute paths, as well as the actual host's cgroup filesystem(s)
to read cgroup metrics.

Since we can't opt out of cgroup namespaces in containers (with k8s
at least), we need a hack to break out of the namespace. For example,
when running in a container with the following mounts:
- host's /proc mounted at /host/proc
- cgroup fs mounted at /cgroup
we observe the following:
- /proc/*/cgroup contains virtualized cgroup names
- /cgroup contains only our cgroup (and its descendants)
- /host/proc/*/cgroup *still* only contains virtualized cgroup names
- /host/proc/1/sys/fs/cgroup (or similar) contains all cgroups

So there's no way to get a process's full cgroup names from within
a cgroup namespace. Thankfully, even the virtualized cgroups can be
converted to full paths with extra context as they're not simply
truncated to / but contain the actual path relative to the cgroup
namespace root, e.g. /../some-other-container or /../../system.slice/...

In order to reconstruct the full cgroup path, we need to know
the absolute path to our own cgroup. Given the above, the only way to
do this is to recursively look for our pid in all the cgroups.
Sadly, this is very expensive, so also provide a way to create
a scap_cgroup_interface without fetching our own cgroup (if you only
need the mounts, for example).

Why do we want a second scap_cgroup_interface? In sinsp we need to
look up cgroup mount directories in a fairly weird context (separate
thread during container metadata lookup without access to `sinsp`
instance). In order to avoid lifetime issues, it's easiest to just
create a second one for lookup_cgroup_dir().

Signed-off-by: Grzegorz Nosek <grzegorz.nosek@sysdig.com>