Initial support for file descriptor lookup
Currently fields such as fd.name refer to the file descriptor used for
the syscall event and it's not possible to refer to other file
descriptors. This makes life difficult when writing rules which need to
evalute the state of multiple file descriptors for e.g. reverse shell
detection.

Improve the situation by introducing fdinfo[fd] which supports all the
existing fd.* fields with and additional fd parameter allowing other
file descriptors to be referenced.

The current implementation is a hack in the parser for fd fields where we
detect the fdinfo string, extract the fd supplied and then rewrite the field as
if it were a regular fd field and send it through the regular parser. A better
solution would be to add an indexable attribute to child classes of
sinsp_filter_check and implement it in the parent class so other fields could
leverage the same functionality (e.g.  proc[pid]).