feat(libsinsp): Allow iterating over file descriptors
Currently the file descriptor table can only be accessed with get_fd() which
is only useful if you know the fd number you want to lookup, it's less useful
if you want to e.g. see if any file descriptors are of a specific type.

Support this use case by providing a way to iterate a callback over all open
file descriptors.

Signed-off-by: Allan Xavier <allan.xavier@sysdig.com>