feat(libsinsp): Support list filters for fd.types
Rework the initial implementation of fd.types so the "fd.types" field
represents a union of all the types used by open file descriptors in the thread
assocaited with the event.

This can then be used with the in and intersects operators for queries such as
'fd.types in ("ipv4", "ipv6")' which will be true if a process has any open
network sockets.

Note this change means fd.types[1] will also now return a list of a single
value instead of just a single value.

Signed-off-by: Allan Xavier <allan.xavier@sysdig.com>