WIP Reparent children during parse_thread_exit.
This duplicates the reparenting calls that remove_thread normally makes. This
is needed when m_automatic_threadtable_purging is set to true as calls to
remove_thread can be much delayed which would result in incorrect process
lineage due to dead threads still being parents of live threads.

The duplication is less than ideal, however given that remove_thread is used in
multiple places it isn't clear whether those other call sites need reparenting
or not and better to be safe than sorry.