Scoped dependencies and generator expressions for libsinsp
In order for a generator expression to correctly select between
OpenSSL and ssl_shim when linking libsinsp to executables,
libsinsp's transitive dependencies needed to be properly scoped.

This commit marks sinsp-internal dependencies as PRIVATE and
transitive third-party dependencies as INTERFACE so that our
generator expressions can prevent conflicts between OpenSSL
and ssl_shim when linking binaries like dragent and agentone.

It also replaces a direct reference to OPENSSL_LIBRARIES with
a generator expression that select OpenSSL or ssl_shim based
on the target.